requirements.txt
SQLAlchemy==2.0.39
psycopg[binary]==3.2.9
python-dotenv==1.0.1